fetch survey with hard coded id

+++ backend/src/Data/Survey/SurveyRowDTO.cs
@@ -4,13 +4,13 @@ public class SurveyRowDTO
 {
     public string Benefit { get; set; }
     public string BenefitType { get; set; }
-    public float Importance { get; set; }
+    public double Importance { get; set; }
     public List<string> Scales { get; set; }
     public string Description { get; set; }
 
     public SurveyRowDTO(string benefit,
         string benefitType,
-        float importance,
+        double importance,
         List<string> scales,
         string description)
     {
